backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
fredapi
scikit-learn
pandas
joblib
python-dotenv

